Added the start to configuring modals for each components settings page.
To create a new modal for a setting page, go after the ButtonConfigOverlayModal.js script. then call the SetConfigOverlayVisable for that specific config modal you created, to true in the onButtonLongPress function within PageId.js. Then in the onButtonPress, you can go through any functionality that should occur when that specific component type is pressed.